Xiaohoz

+++ playlist/remove_media.php
@@ -1,6 +1,10 @@
 <?php
 session_start();
 include_once "../function.php";
+if(!$_SESSION["username"]){
+  header("Location: ../login_register/login.php");
+  exit();
+}
 $playlistid = $_GET['playlistid'];
 $mediaid = $_GET['mediaid'];
 $result = mysql_query("delete from playlistmedia where playlistid='$playlistid' and mediaid='$mediaid'")
